+++ src/capabilities/get-events-for-cluster.ts
@@ -9,5 +9,5 @@ export const getEventsForCluster = async (dtClient: HttpClient, clusterId: strin
     dql = `fetch events | filter isNotNull(k8s.cluster.uid)`;
   }
 
-  return executeDql(dtClient, dql);
+  return executeDql(dtClient, { query: dql });
 };